refactor(superblock): move superblock creation out of mining
